move Payment component tests to Payment/__tests__ and rename SendTab to PaymentTab.
