+++ client/src/components/pages/Store/Wallet/Transactions/Payment/__tests__/PaymentTab.test.jsx
@@ -5,7 +5,7 @@ import BitcoinPriceService from "@/services/bitcoinPriceService";
 import * as walletService from "@/services/walletService";
 import { I18nProvider } from "@i18n/I18nProvider";
 
-import { PaymentTab } from "../Payment";
+import { PaymentTab } from "..";
 
 jest.mock("@heroui/react", () => {
   const actual = jest.requireActual("@heroui/react");
@@ -211,7 +211,7 @@ describe("SendTab Component", () => {
       fireEvent.click(screen.getByText("payments.send.confirmModal.confirmButton"));
 
       await waitFor(() => {
-        expect(walletService.payInvoiceFromService).toHaveBeenCalledWith("lnbc1000n1pj9h8uqpp5test", null);
+        expect(walletService.payInvoiceFromService).toHaveBeenCalledWith("lnbc1000n1pj9h8uqpp5test", null, expect.any(Object));
       });
     });
 
@@ -349,7 +349,7 @@ describe("SendTab Component", () => {
       fireEvent.click(screen.getByText("payments.send.confirmModal.confirmButton"));
 
       await waitFor(() => {
-        expect(walletService.payInvoiceFromService).toHaveBeenCalledWith("lnbc1000n1pj9h8uqpp5test", 5000);
+        expect(walletService.payInvoiceFromService).toHaveBeenCalledWith("lnbc1000n1pj9h8uqpp5test", 5000, expect.any(Object));
       });
     });
 
@@ -379,7 +379,7 @@ describe("SendTab Component", () => {
       fireEvent.click(screen.getByText("payments.send.confirmModal.confirmButton"));
 
       await waitFor(() => {
-        expect(walletService.payInvoiceFromService).toHaveBeenCalledWith("lnbc1000n1pj9h8uqpp5test", 5000);
+        expect(walletService.payInvoiceFromService).toHaveBeenCalledWith("lnbc1000n1pj9h8uqpp5test", 5000, expect.any(Object));
       });
     });
   });
