+++ client/src/components/pages/Store/Wallet/Transactions/Payment/__tests__/ZeroAmountPaymentFields.test.jsx
@@ -2,7 +2,7 @@ import { render, screen, fireEvent } from "@testing-library/react";
 
 import { I18nProvider } from "@i18n/I18nProvider";
 
-import { ZeroAmountPaymentFields } from "../Payment/ZeroAmountPaymentFields";
+import { ZeroAmountPaymentFields } from "../ZeroAmountPaymentFields";
 
 jest.mock("@/components/hooks/useCurrency", () => ({
   useCurrency: () => ({
